+++ src/add.c
@@ -10,7 +10,7 @@
 // given by s1 and s2, and stores its length at rlen
 // requires: s1 and s2 represent Zeckendorf representations of length len with possible leading ZEROs, rlen != NULL
 // effects: allocates memory (caller must free), updates *rlen
-static char *add_same_len(const char *s1, const char *s2, const size_t len, size_t *rlen) {
+static char *add_same_len(const char *s1, const char *s2, const int len, int *rlen) {
 	const char TWO = ONE + 1;
 	const char THREE = TWO + 1;
 
@@ -118,7 +118,7 @@ static char *add_same_len(const char *s1, const char *s2, const size_t len, size
 	return ans;
 }
 
-char *add_len(const char *s1, const char *s2, const size_t len1, const size_t len2, size_t *rlen) {
+char *add_len(const char *s1, const char *s2, const int len1, const int len2, int *rlen) {
 	// add leading ZEROs to make lengths equal, then use add_same_len
 	if (len1 > len2) {
 		char *cp = malloc(len1 * sizeof(char));
